Complete rewrite of the position saving system. We now save based on Left and Top which are more natural numbers.
This allows us to use the Left / Top coordinates that we get when a Toast is dragged around the screen, as well as better off-screen detection.

DPI is only used when calculating start positions (i.e. first run / Settings reset events etc) - see ScreenHelper.cs::GetDefaultToastPosition

TODO: this needs verification on a mixed DPI system

BREAKING CHANGE: This will force a reset of all user's existing Toast locations. Since we have Ctrl+Drag to reposition Toast this is not considered a serious issue