PostBuild was ignoring chromedriver copy step
To override the non-0 return value of taskkill if chromedriver wasn't
running we were forcing an exit 0, but that prevented the copy if it
wasn't running.

It is safe to remove this since the returned code will be the copy code,
which will succeed (unless there was a real error) if chromedriver is
not executing.