Pass around Song objects instead of strings
We were representing songs as a string formatted as "Artist - Song" and
then parsing that out in various places (using string parsing).

This broke various scenarios (especially when the Artist or song
included a -) and was wasteful. Instead we now pass around Song objects
which have both Artist and Title as properties as well as a ToString()
for outputting the formatted version if needed.