Minimize Toast when not in use
Due to the amount of time it takes to start Spotify with the new
infrastructure we need to start Minimized. As part of this also added
code to minimize the tray when not actually showed, this removes the
transparent always topmost window which causes FPS drops in demanding
games (such as WoW) on some GPUs.