Remove references to ChromeDriver
Now that Spotify have the currently playing song back in the title, we
can remove ChromeDriver.

While this removes some of the cooler functionality that we had in
Toastify, it was also causing slowness as well as causing Spotify to
consume a lot more resources than it should have, so we're reverting to
the old behaviour.