Ensure that album art is always preserved
Our existing behaviour calls for retrieving the current playing song on
every hotkey action (actually, it was being retrieved twice), but with
the new behaviour of storing the album art inside the Song class we
ended up with a null url (since those calling points were not setting
the album art like CheckTitle() does.

To work around this, and improve performance while we're at it, we no
longer recheck the current song (since there is no benefit to doing so,
we already know what the current song is) and rely on it for setting the
toast icon.

Note: renamed previousTitle to currentSong to more accurately reflect
its purpose.