Fixed error on start-up (#3)
The *"Object reference not set to an instance of an object"* error was caused by Spotify not being completely loaded when Toastify tried to get its status.
Actually, this was caused by a dumb null reference in Song.cs; however, the fact that Toastify started before Spotify was fully initialized, caused some unidentified problems in the connection enstablished by the local APIs which in turn were the cause of issue #1.