Changed IsForegroundAppRunningInFullscreen
Checking for WS_MAXIMIZE and WS_POPUP was not enough, and more importantly not correct.
For the moment, if a window is WS_EX_TOPMOST and its size is equal to the screen size, then it is considered fullscreen (as in 'exclusive fullscreen' and not 'windowed borderless') and the toast will not fade-in, causing the app window to lose focus and go to the background/minimize.