Fixed Toast behaviour with fullscreen apps (specially videogames)
The problem of the Toast stealing the focus from fullscreen videogames should be definitely fixed.
The problem happened at the end of the fade-out event. Previously, this was fixed by preventing the fade-in in case one of these apps was detected; however, recent tests with several videogames showed that it was unreliable, because every application can implement its 'fullscreen' and 'borderless windowed' modes differently (see previous commit).
Now, we just prevent the Toast to became 'active' and stealing the focus when it is minimized at the end of the fade-out event (see the new ShowToast method in Toast.xaml.cs).