Now StartSpotify runs in a background thread
If an exception is fired or Toastify is not able to connect with Spotify within a time limit (1 minute), the background worker terminates Toastify with an error message, asking the user to restart Toastify after having started Spotify manually.
Also, while StartSpotify is running, the default tray icon is replaced by a *loading* animated icon; while loading, the Settings window will be unaccessible and the hotkeys won't work.

This should fix #13.