Fix some PInvoke's to account for platform bitness (#27)
If Toastify was built for AnyCPU and run on a 32bit OS, or built for AnyCPU (Prefer 32bit) and run on a 64bit OS, it would run as a 32bit process and load 32bit libraries.
Specifically, the problem was that the 32bit version of user32.dll doesn't contain the 'GetWindowLongPtr' and 'SetWindowLongPtr' functions and Toastify would crash on startup.