Switch to Json serialization
The change from XML to JSON serialization is driven from the need of being able to populate an already available object with serialized data (#30).
The Newtonsoft Json.NET library, which is already used by one of the dependencies of this project (so, there are no new dependencies), provides this functionality.