Upgrade to .NET 4.7.2
It has been necessary to upgrade to .NET 4.7.2 because of the transitive dependencies from "Aleab.Common.Net.WebSockets". Being a NetStandard 2.0 library, that submodule imports loads and loads of other dependencies in .NET Framework versions <4.7.2 (notably from the System.* namespace).

By upgrading to 4.7.2, the majority of these dependencies are not needed anymore; some of them are still necessary though (some from System.*, most from Microsoft.*).
The number of extra dependencies has been reduced from ~200 to 44.

All the dependencies have been moved to the /lib subfolder of the installation directory.

===

In addition to the changes above, the project has been fixed to correctly compile the WPF items.